requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pillow==10.1.0
opencv-python-headless==4.8.1.78
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
ultralytics==8.0.196
python-multipart==0.0.6
accelerate==0.24.1
huggingface-hub==0.19.4
